+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.6] - 2021-07-20
+### Added
+- 디버그 로그용 클래스 및 매서드 Printer.Print() 추가.
+- ColorExtension에 Darken 및 Gray 추가.
+
 ## [1.1.5] - 2021-07-16
 ### Added
 - 간단한 SerializableDictionary 만듬 (OcDictionary)